更新老版资源链接

+++ postbuild.sh
@@ -1,7 +1,7 @@
 ts=$(date +%s) 
-echo '{\"version\":'$tmp'}'>./dist/aktools/version.json
-echo '{\"version\":'$tmp'}'>./dist/aktools-old/version.json
-#     "postbuild": "mv ./dist/my-project/*.js ./dist/my-project/*.js.map ./dist/my-project/resources; sed -i -e 's/src=\"/src=\"resources\\//g' dist/my-project/index.html",
+echo '{"version":'$ts'}'>./dist/aktools/version.json
+echo '{"version":'$ts'}'>./dist/aktools-old/version.json
+
 mkdir ./dist/aktools/static/js
 mkdir ./dist/aktools/static/css
 
@@ -13,6 +13,16 @@ sed -i -e 's#styles.css#static/aktools/'$ts'/css/styles.css#g' ./dist/aktools/in
 
 sed -i -e 's#</noscript>#</noscript><script>window.data_version='$ts';</script>#g' ./dist/aktools/index.html
 
-rm ./dist/aktools/index.html-e
+
+[ -d "./dist/aktools-old/" ] && rm -rf ./dist/aktools-old
+cp -R ./aktools-old ./dist/
+
+sed -i -e 's#"static/#static/aktools-old/'$ts'/' ./dist/aktools-old/akchars.html
+sed -i -e 's#"static/#static/aktools-old/'$ts'/' ./dist/aktools-old/akevolve.html
+sed -i -e 's#"static/#static/aktools-old/'$ts'/' ./dist/aktools-old/akhr.html
+sed -i -e 's#"static/#static/aktools-old/'$ts'/' ./dist/aktools-old/akhrchars.html
+sed -i -e 's#"static/#static/aktools-old/'$ts'/' ./dist/aktools-old/aklevel.html
+
+[ -f "./dist/aktools/index.html-e"] && rm ./dist/aktools/index.html-e
 
 cp ./dist/aktools/index.html ./dist/aktools/404.html
